% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity_fn.R
\name{bw_MSE}
\alias{bw_MSE}
\title{Internal function.}
\usage{
bw_MSE(data, grid, p, v, kernel, Cweights, Pweights, regularize)
}
\arguments{
\item{data, }{Numeric vector, the data.}

\item{grid, }{Numeric vector, the evaluation points.}

\item{p, }{Integer, polynomial order.}

\item{v, }{Integer, order of derivative.}

\item{kernel, }{String, the kernel.}

\item{Cweights, }{Numeric vector, the counterfactual weights.}

\item{Pweights, }{Numeric vector, the survey sampling weights.}

\item{Regularize, }{whether the bandwidth should be regularized.}
}
\value{
\item{}{Bandwidth sequence.}
}
\description{
Calculates MSE-optimal bandwidths.
}
\keyword{internal}
