% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity_fn.R
\name{bw_ROT}
\alias{bw_ROT}
\title{Internal function.}
\usage{
bw_ROT(data, grid, p, v, kernel, regularize)
}
\arguments{
\item{data, }{Numeric vector, the data.}

\item{grid, }{Numeric vector, the evaluation points.}

\item{p, }{Integer, polynomial order.}

\item{v, }{Integer, order of derivative.}

\item{kernel, }{String, the kernel.}

\item{Regularize, }{whether the bandwidth should be regularized.}
}
\value{
\item{}{Bandwidth sequence.}
}
\description{
Calculates rule-of-thumb bandwidth
}
\keyword{internal}
