% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity_fn.R
\name{Cgenerate}
\alias{Cgenerate}
\title{Internal function.}
\usage{
Cgenerate(k, p, low = -1, up = 1, kernel = "triangular")
}
\arguments{
\item{k}{Nonnegative integer, extra order (usually p+1).}

\item{p}{Nonnegative integer, the polynomial order.}

\item{low, up}{Scalar, between -1 and 1, region of integration.}

\item{kernel}{String, the kernel function.}
}
\value{
A (p+1)-by-1 matrix.
}
\description{
Generate matrix.
}
\keyword{internal}
