% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity_fn.R
\name{normal_dgps}
\alias{normal_dgps}
\title{Internal function.}
\usage{
normal_dgps(x, v, mean, sd)
}
\arguments{
\item{x, }{Numeric, point of evaluation.}

\item{v, }{Numeric and nonnegative. The derivative, with 0 being cdf and 1 being pdf.'}

\item{mean, sd, }{the mean and standard deviation.}
}
\value{
\item{}{A scalar.}
}
\description{
Calculates density and higher order derivatives for Gaussian models.
}
\keyword{internal}
