% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity.R
\name{print.CJMlpbwdensity}
\alias{print.CJMlpbwdensity}
\title{Internal function.}
\usage{
\method{print}{CJMlpbwdensity}(x, ...)
}
\arguments{
\item{x}{Class \code{CJMlpbwdensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
