% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdensity.R
\name{print.CJMlpdensity}
\alias{print.CJMlpdensity}
\title{Internal function.}
\usage{
\method{print}{CJMlpdensity}(x, ...)
}
\arguments{
\item{x}{Class \code{CJMlpdensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
