% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdensity_methods.R
\name{print.lpdensity}
\alias{print.lpdensity}
\title{Print Method for Local Polynomial Density Estimation and Inference}
\usage{
\method{print}{lpdensity}(x, ...)
}
\arguments{
\item{x}{Class "lpdensity" object, obtained from calling \code{\link{lpdensity}}.}

\item{...}{Additional options.}
}
\description{
The print method for local polynomial density objects.
}
\examples{
# Generate a random sample
set.seed(42); X <- rnorm(2000)

# Estimate density and report results
print(lpdensity(data = X, bwselect = "imse-dpi"))

}
\seealso{
\code{\link{lpdensity}} for local polynomial density estimation.

Supported methods: \code{\link{coef.lpdensity}}, \code{\link{confint.lpdensity}},
  \code{\link{plot.lpdensity}}, \code{\link{print.lpdensity}}, \code{\link{summary.lpdensity}},
  \code{\link{vcov.lpdensity}}.
}
\author{
Matias D. Cattaneo (maintainer), Princeton University. \email{matias.d.cattaneo@gmail.com}.

Michael Jansson, University of California Berkeley. \email{michael.jansson.berkeley@gmail.com}.

Xinwei Ma, University of California San Diego. \email{xinweima.pku@gmail.com}.
}
