% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity_fn.R
\name{Sgenerate}
\alias{Sgenerate}
\title{Internal function.}
\usage{
Sgenerate(p, low = -1, up = 1, kernel = "triangular")
}
\arguments{
\item{p}{Nonnegative integer, polynomial order.}

\item{low, up}{Scalar, between -1 and 1, the region of integration.}

\item{kernel}{String, the kernel function.}
}
\value{
A (p+1)-by-(p+1) matrix.
}
\description{
Generate matrix.
}
\keyword{internal}
