% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity.R
\name{summary.CJMlpbwdensity}
\alias{summary.CJMlpbwdensity}
\title{Internal function.}
\usage{
\method{summary}{CJMlpbwdensity}(object, ...)
}
\arguments{
\item{object}{Class \code{CJMlpbwdensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
