% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdensity.R
\name{summary.CJMlpdensity}
\alias{summary.CJMlpdensity}
\title{Internal function.}
\usage{
\method{summary}{CJMlpdensity}(object, ...)
}
\arguments{
\item{object}{Class \code{CJMlpdensity} objects.}
}
\description{
Internal function.
}
\keyword{internal}
