% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpbwdensity_fn.R
\name{Tgenerate}
\alias{Tgenerate}
\title{Internal function.}
\usage{
Tgenerate(p, low = -1, up = 1, kernel = "triangular")
}
\arguments{
\item{p, }{Integer, polynomial order.}

\item{low, up, }{Numeric, between -1 and 1, region of integration.}

\item{kernel, }{String, the kernel.}
}
\value{
\item{}{A (p+1)-by-(p+1) matrix.}
}
\description{
Generate matrix.
}
\keyword{internal}
